Renamed ClassLoaderAction to StorageAction and moved to org.apache.kafka.server.log.internals package in storage module
